requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
python-dotenv==1.0.0